requirements.txt
streamlit==1.32.0
torch==2.3.0
torchvision==0.18.0
transformers==4.44.0
Pillow==10.3.0
numpy==1.26.4
google-generativeai
python-dotenv==1.0.1